change interface on patch_okuribito
